# Conflicts:
#	csir_soil_sensor/lib/src/data/db/app_database.dart
#	csir_soil_sensor/lib/src/data/db/app_database.g.dart
#	csir_soil_sensor/lib/src/features/crop_params/crop_form_screen.dart
#	csir_soil_sensor/lib/src/services/bluetooth_service.dart
#	csir_soil_sensor/lib/src/services/export_service.dart
#	csir_soil_sensor/test/db_repository_test.dart

+++ csir_soil_sensor/lib/src/services/bluetooth_service.dart
@@ -747,7 +747,7 @@ class BluetoothService extends StateNotifier<BluetoothStateModel> {
             phosphorus: reading.phosphorus,
             potassium: reading.potassium,
             salinity: reading.salinity,
-            tds: drift.Value(reading.tds),
+            tds: drift.Value(reading.tds.toDouble()),
             ecConv: drift.Value(reading.ecConv),
             ecCal: drift.Value(reading.ecCal),
             phConv: drift.Value(reading.phConv),
